Module Syntax.SymVar

Identifier of symbolically represented variable

include ID

type t

val create : string -> TI.t -> t

val get_name : t -> string

val get_ti : t -> TI.t

val add_array_index : t -> int -> t

add_array_index var index Add array addressation to index. There may be more than one indexes for multi-dimension arrays, e.g. A13.

val add_array_index_opaque : t -> t

add_array_index var index Add array addressation to index. There may be more than one indexes for multi-dimension arrays, e.g. A13.

add_array_index_opaque var Add array addressation to unknown index that can't be evaluated before run time, e.g. Af(a,b).

val get_array_indexes : t -> int option list

add_array_index_opaque var Add array addressation to unknown index that can't be evaluated before run time, e.g. Af(a,b).

val to_yojson : t -> Yojson.Safe.t