Module Syntax.TimeValue

Representation of time interval

According the IEC61131-3 3rd edition grammar, all interval values defined as fix_point. That means that these values could be represented as float values.

type t

val mk : ?y:int -> ?mo:int -> ?d:float -> ?h:float -> ?m:float -> ?s:float -> ?ms:float -> ?us:float -> ?ns:float -> unit -> t

val (+) : t -> t -> t

val inv : t -> t

Invert time value to represent a negative duration.

val to_string : t -> string

val is_zero : t -> bool

val to_yojson : t -> Yojson.Safe.t